Venti.
Untested, use with caution.
For test with Rafa's fossil work.

+++ sys/src/cmd/venti/srv/whack.c
@@ -136,7 +136,7 @@ whackmatch(Whack *b, uint8_t **ss, uint8_t *esrc, uint32_t h, uint32_t now)
  * and might be faster on some machines
  */
 /*
-#define hashit(c)	((((ulong)(c) * 0x6b43a9) >> (24 - HashLog)) & HashMask)
+#define hashit(c)	((((uint32_t)(c) * 0x6b43a9) >> (24 - HashLog)) & HashMask)
 */
 #define hashit(c)	(((((uint32_t)(c) & 0xffffff) * 0x6b43a9b5) >> (32 - HashLog)) & HashMask)
